pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ichtaka"
version = "0.1.0"
description = "Anonymous reporting platform."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.3",
    "fastapi>=0.121.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[project.scripts]
start = "src.app:start"

[tool.setuptools]
packages = ["src"]

[tool.uv]
package = true
